src/requirements.txt
Flask==3.0.3
numpy==1.26.4
opencv_python==4.9.0.80
pandas==2.2.2
PyYAML==6.0.1
PyYAML==6.0.1
Requests==2.32.1
scikit_learn==1.4.0
tensorflow==2.16.1
tensorflow_intel==2.16.1
ultralytics==8.2.16
